Add function to tables and map controllers in CoreService
