+++ CoreService/Data/TerranovaDbContext.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using CoreService.Model.Terranova;
+using CoreService.Models.Terranova;
 using Microsoft.EntityFrameworkCore;
 
 namespace CoreService.Data
