+++ CoreService/Models/Terranova/OrderItem.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using System.ComponentModel.DataAnnotations;
 
-namespace CoreService.Model.Terranova
+namespace CoreService.Models.Terranova
 {
     public class OrderItem
     {
