+++ Database/TerranovaDatabase/README.md
@@ -5,6 +5,7 @@ This repository contains all information you need to create and know the design 
 ## Contents
 
 - Database creation script
+- User information
 - Structure
 	-Tables with relationships and foreign keys
 	- Stored procedures
@@ -20,6 +21,10 @@ This repository contains all information you need to create and know the design 
 
 1. Open SQL Server Management Studio.
 2. Open Data folder and run the scripts in order.
+
+## User information
+- User -> terranova
+- password -> 123456789
 
 ## Structure
